TraceRasterizer DI for Rasterizer
